+++ src/entities/preference/index.ts
@@ -1,0 +1,1 @@
+export { usePrefsStore } from '#/entities/preference/model/store'